:sparkles: metadata functions and presentation in yaml
